requirements.txt
django-instant>=0.6
goerr